+++ src/anima/commands/__init__.py
@@ -1,0 +1,8 @@
+from anima.commands import attack
+from anima.commands.combo import combo
+from anima.commands.crit import crit
+from anima.commands.mass import mass
+from anima.commands.powers import powers
+from anima.commands.shuffle import shuffle
+
+__all__ = ["attack", "powers", "combo", "crit", "mass", "shuffle"]